chore(internal): move browser commands from folders to files
