+++ lib/browser/commands/getConfig.js
