+++ lib/browser/commands/index.js
@@ -2,7 +2,7 @@
 
 module.exports = [
     'assert-view',
-    'set-orientation',
-    'get-config',
-    'getPuppeteer'
+    'getConfig',
+    'getPuppeteer',
+    'setOrientation'
 ];
