+++ lib/browser/commands/setOrientation.js
